Address review stage 1 (part 2) of PR 55657
